Upload files

+++ src/main/java/com/rayferric/comet/scenegraph/node/Node.java
@@ -1,7 +1,7 @@
 package com.rayferric.comet.scenegraph.node;
 
-import com.rayferric.comet.math.Matrix4d;
-import com.rayferric.comet.math.Vector3d;
+import com.rayferric.comet.math.Matrix4f;
+import com.rayferric.comet.math.Vector3f;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,9 +12,9 @@ public class Node {
     public Node() {
         name.set("Node");
 
-        setTranslation(new Vector3d(0));
-        setRotation(new Vector3d(0));
-        setScale(new Vector3d(1));
+        setTranslation(new Vector3f(0));
+        setRotation(new Vector3f(0));
+        setScale(new Vector3f(1));
 
         updateLocalTransform();
     }
@@ -112,48 +112,48 @@ public class Node {
         }
     }
 
-    public Vector3d getTranslation() {
-        return new Vector3d(translation.get());
+    public Vector3f getTranslation() {
+        return new Vector3f(translation.get());
     }
 
-    public void setTranslation(Vector3d translation) {
-        this.translation.set(new Vector3d(translation));
+    public void setTranslation(Vector3f translation) {
+        this.translation.set(new Vector3f(translation));
         localTransformValid.set(false);
         invalidateGlobalTransform();
     }
 
-    public Vector3d getRotation() {
-        return new Vector3d(rotation.get());
+    public Vector3f getRotation() {
+        return new Vector3f(rotation.get());
     }
 
-    public void setRotation(Vector3d rotation) {
-        this.rotation.set(new Vector3d(rotation));
+    public void setRotation(Vector3f rotation) {
+        this.rotation.set(new Vector3f(rotation));
         localTransformValid.set(false);
         invalidateGlobalTransform();
     }
 
-    public Vector3d getScale() {
-        return new Vector3d(scale.get());
+    public Vector3f getScale() {
+        return new Vector3f(scale.get());
     }
 
-    public void setScale(Vector3d scale) {
-        this.scale.set(new Vector3d(scale));
+    public void setScale(Vector3f scale) {
+        this.scale.set(new Vector3f(scale));
         localTransformValid.set(false);
         invalidateGlobalTransform();
     }
 
-    public Matrix4d getLocalTransform() {
+    public Matrix4f getLocalTransform() {
         synchronized(localTransformValid) {
             if(!localTransformValid.get()) updateLocalTransform();
         }
-        return new Matrix4d(localTransformCache);
+        return new Matrix4f(localTransformCache);
     }
 
-    public Matrix4d getGlobalTransform() {
+    public Matrix4f getGlobalTransform() {
         synchronized(globalTransformValid) {
             if(!globalTransformValid.get()) updateGlobalTransform();
         }
-        return new Matrix4d(globalTransformCache);
+        return new Matrix4f(globalTransformCache);
     }
 
     private final AtomicReference<String> name = new AtomicReference<>();
@@ -164,20 +164,20 @@ public class Node {
     private final Object parentLock = new Object();
     private final Object childrenLock = new Object();
 
-    private final AtomicReference<Vector3d> translation = new AtomicReference<>();
-    private final AtomicReference<Vector3d> rotation = new AtomicReference<>();
-    private final AtomicReference<Vector3d> scale = new AtomicReference<>();
+    private final AtomicReference<Vector3f> translation = new AtomicReference<>();
+    private final AtomicReference<Vector3f> rotation = new AtomicReference<>();
+    private final AtomicReference<Vector3f> scale = new AtomicReference<>();
 
-    private Matrix4d localTransformCache;
+    private Matrix4f localTransformCache;
     private final AtomicBoolean localTransformValid = new AtomicBoolean();
     // localTransformValid is also the sync object
 
-    private Matrix4d globalTransformCache;
+    private Matrix4f globalTransformCache;
     private final AtomicBoolean globalTransformValid = new AtomicBoolean();
     // globalTransformValid is also the sync object
 
     private void updateLocalTransform() {
-        localTransformCache = Matrix4d.transform(translation.get(), rotation.get(), scale.get());
+        localTransformCache = Matrix4f.transform(translation.get(), rotation.get(), scale.get());
         localTransformValid.set(true);
     }
